Refactor events
CHANGELOG_BEGIN
CHANGELOG_END
