+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/store/dao/JdbcLedgerDao.scala
@@ -50,7 +50,7 @@ import com.digitalasset.platform.events.EventIdFormatter.split
 import com.digitalasset.platform.store.Contract.{ActiveContract, DivulgedContract}
 import com.digitalasset.platform.store.Conversions._
 import com.digitalasset.platform.store.dao.JdbcLedgerDao.{H2DatabaseQueries, PostgresQueries}
-import com.digitalasset.platform.store.dao.events.{TransactionsReader, TransactionWriter}
+import com.digitalasset.platform.store.dao.events.{TransactionsReader, TransactionsWriter}
 import com.digitalasset.platform.store.entries.LedgerEntry.Transaction
 import com.digitalasset.platform.store.entries.{
   ConfigurationEntry,
@@ -1735,8 +1735,8 @@ private class JdbcLedgerDao(
       ()
     }
 
-  override val transactionsWriter: TransactionWriter =
-    TransactionWriter
+  override val transactionsWriter: TransactionsWriter =
+    TransactionsWriter
 
   override val transactionsReader: TransactionsReader =
     TransactionsReader(dbDispatcher, dbType, executionContext)
