+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/store/dao/events/EventsTable.scala
@@ -6,26 +6,22 @@ package com.digitalasset.platform.store.dao.events
 import java.io.InputStream
 import java.util.Date
 
-import anorm.{
-  BatchSql,
-  NamedParameter,
-  Row,
-  RowParser,
-  SimpleSql,
-  SqlParser,
-  SqlStringInterpolation,
-  ~
-}
+import anorm.SqlParser._
+import anorm.{RowParser, ~}
 import com.daml.ledger.participant.state.v1.Offset
 import com.digitalasset.daml.lf.data.Ref.QualifiedName
 import com.digitalasset.ledger.api.v1.event.{ArchivedEvent, CreatedEvent, Event}
+import com.digitalasset.ledger.api.v1.transaction.{Transaction => ApiTransaction}
+import com.digitalasset.ledger.api.v1.transaction_service.GetFlatTransactionResponse
 import com.digitalasset.ledger.api.v1.value.Identifier
-import com.digitalasset.ledger.{ApplicationId, CommandId, TransactionId, WorkflowId}
-import com.digitalasset.platform.events.EventIdFormatter.fromTransactionId
+import com.digitalasset.platform.ApiOffset
+import com.digitalasset.platform.index.TransactionConversion
 import com.digitalasset.platform.participant.util.LfEngineToApi
-import com.digitalasset.platform.store.Conversions._
-import com.digitalasset.platform.store.serialization.ValueSerializer
-import com.digitalasset.platform.store.serialization.ValueSerializer.{serializeValue => serialize}
+import com.digitalasset.platform.store.Conversions.offset
+import com.digitalasset.platform.store.serialization.ValueSerializer.{
+  deserializeValue => deserialize
+}
+import com.google.protobuf.timestamp.Timestamp
 
 /**
   * Data access object for a table representing raw transactions nodes that
@@ -33,459 +29,147 @@ import com.digitalasset.platform.store.serialization.ValueSerializer.{serializeV
   * with a [[WitnessesTable]] events can be filtered based on their visibility to
   * a party.
   */
-private[events] object EventsTable {
+private[events] object EventsTable
+    extends EventsTable
+    with EventsTableInsert
+    with EventsTableFlatEvents
 
-  private def cantSerialize(attribute: String, forContract: ContractId): String =
-    s"Cannot serialize $attribute for ${forContract.coid}"
+private[events] trait EventsTable {
 
-  private def serializeCreateArgOrThrow(node: Create): Array[Byte] =
-    serialize(
-      value = node.coinst.arg,
-      errorContext = cantSerialize(attribute = "create argument", forContract = node.coid),
-    )
-
-  private def serializeNullableKeyOrThrow(node: Create): Option[Array[Byte]] =
-    node.key.map(
-      k =>
-        serialize(
-          value = k.key,
-          errorContext = cantSerialize(attribute = "key", forContract = node.coid),
-      ))
-
-  private def serializeExerciseArgOrThrow(node: Exercise): Array[Byte] =
-    serialize(
-      value = node.chosenValue,
-      errorContext = cantSerialize(attribute = "exercise argument", forContract = node.targetCoid),
-    )
-
-  private def serializeNullableExerciseResultOrThrow(node: Exercise): Option[Array[Byte]] =
-    node.exerciseResult.map(exerciseResult =>
-      serialize(
-        value = exerciseResult,
-        errorContext = cantSerialize(attribute = "exercise result", forContract = node.targetCoid),
-    ))
-
-  private val insertCreate: String =
-    """insert into participant_events(
-          |  event_id,
-          |  event_offset,
-          |  contract_id,
-          |  transaction_id,
-          |  workflow_id,
-          |  ledger_effective_time,
-          |  template_package_id,
-          |  template_name,
-          |  node_index,
-          |  is_root,
-          |  command_id,
-          |  application_id,
-          |  submitter,
-          |  create_argument,
-          |  create_signatories,
-          |  create_observers,
-          |  create_agreement_text,
-          |  create_consumed_at,
-          |  create_key_value
-          |) values (
-          |  {event_id},
-          |  {event_offset},
-          |  {contract_id},
-          |  {transaction_id},
-          |  {workflow_id},
-          |  {ledger_effective_time},
-          |  {template_package_id},
-          |  {template_name},
-          |  {node_index},
-          |  {is_root},
-          |  {command_id},
-          |  {application_id},
-          |  {submitter},
-          |  {create_argument},
-          |  {create_signatories},
-          |  {create_observers},
-          |  {create_agreement_text},
-          |  null,
-          |  {create_key_value}
-          |)
-          |""".stripMargin
-
-  private def create(
-      applicationId: Option[ApplicationId],
-      workflowId: Option[WorkflowId],
-      commandId: Option[CommandId],
-      transactionId: TransactionId,
-      nodeId: NodeId,
-      submitter: Option[Party],
-      roots: Set[NodeId],
+  case class Entry[E](
+      eventOffset: Offset,
+      transactionId: String,
       ledgerEffectiveTime: Date,
-      offset: Offset,
-      create: Create,
-  ): Vector[NamedParameter] =
-    Vector[NamedParameter](
-      "event_id" -> fromTransactionId(transactionId, nodeId).toString,
-      "event_offset" -> offset,
-      "contract_id" -> create.coid.coid.toString,
-      "transaction_id" -> transactionId.toString,
-      "workflow_id" -> workflowId.map(_.toString),
-      "ledger_effective_time" -> ledgerEffectiveTime,
-      "template_package_id" -> create.coinst.template.packageId.toString,
-      "template_name" -> create.coinst.template.qualifiedName.toString,
-      "node_index" -> nodeId.index,
-      "is_root" -> roots(nodeId),
-      "command_id" -> commandId.map(_.toString),
-      "application_id" -> applicationId.map(_.toString),
-      "submitter" -> submitter.map(_.toString),
-      "create_argument" -> serializeCreateArgOrThrow(create),
-      "create_signatories" -> create.signatories.map(_.toString).toArray,
-      "create_observers" -> create.stakeholders.diff(create.signatories).map(_.toString).toArray,
-      "create_agreement_text" -> Some(create.coinst.agreementText).filter(_.nonEmpty),
-      "create_key_value" -> serializeNullableKeyOrThrow(create),
-    )
+      commandId: String,
+      workflowId: String,
+      event: E,
+  )
 
-  private val insertExercise =
-    """insert into participant_events(
-          |  event_id,
-          |  event_offset,
-          |  contract_id,
-          |  transaction_id,
-          |  workflow_id,
-          |  ledger_effective_time,
-          |  template_package_id,
-          |  template_name,
-          |  node_index,
-          |  is_root,
-          |  command_id,
-          |  application_id,
-          |  submitter,
-          |  exercise_consuming,
-          |  exercise_choice,
-          |  exercise_argument,
-          |  exercise_result,
-          |  exercise_actors,
-          |  exercise_child_event_ids
-          |) values (
-          |  {event_id},
-          |  {event_offset},
-          |  {contract_id},
-          |  {transaction_id},
-          |  {workflow_id},
-          |  {ledger_effective_time},
-          |  {template_package_id},
-          |  {template_name},
-          |  {node_index},
-          |  {is_root},
-          |  {command_id},
-          |  {application_id},
-          |  {submitter},
-          |  {exercise_consuming},
-          |  {exercise_choice},
-          |  {exercise_argument},
-          |  {exercise_result},
-          |  {exercise_actors},
-          |  {exercise_child_event_ids}
-          |)
-          |""".stripMargin
+  object Entry {
 
-  private def exercise(
-      applicationId: Option[ApplicationId],
-      workflowId: Option[WorkflowId],
-      commandId: Option[CommandId],
-      transactionId: TransactionId,
-      nodeId: NodeId,
-      submitter: Option[Party],
-      roots: Set[NodeId],
-      ledgerEffectiveTime: Date,
-      offset: Offset,
-      exercise: Exercise,
-  ): Vector[NamedParameter] =
-    Vector[NamedParameter](
-      "event_id" -> fromTransactionId(transactionId, nodeId).toString,
-      "event_offset" -> offset,
-      "contract_id" -> exercise.targetCoid.coid.toString,
-      "transaction_id" -> transactionId.toString,
-      "workflow_id" -> workflowId.map(_.toString),
-      "ledger_effective_time" -> ledgerEffectiveTime,
-      "template_package_id" -> exercise.templateId.packageId.toString,
-      "template_name" -> exercise.templateId.qualifiedName.toString,
-      "node_index" -> nodeId.index,
-      "is_root" -> roots(nodeId),
-      "command_id" -> commandId.map(_.toString),
-      "application_id" -> applicationId.map(_.toString),
-      "submitter" -> submitter.map(_.toString),
-      "exercise_consuming" -> exercise.consuming,
-      "exercise_choice" -> exercise.choiceId.toString,
-      "exercise_argument" -> serializeExerciseArgOrThrow(exercise),
-      "exercise_result" -> serializeNullableExerciseResultOrThrow(exercise),
-      "exercise_actors" -> exercise.actingParties.map(_.toString).toArray,
-      "exercise_child_event_ids" -> exercise.children
-        .map(fromTransactionId(transactionId, _): String)
-        .toArray,
-    )
-
-  private val updateArchived =
-    """update participant_events set create_consumed_at={consumed_at} where contract_id={contract_id}"""
-
-  private def archive(
-      contractId: ContractId,
-      consumedAt: Offset,
-  ): Vector[NamedParameter] =
-    Vector[NamedParameter](
-      "consumed_at" -> consumedAt,
-      "contract_id" -> contractId.coid.toString,
-    )
-
-  sealed abstract case class PreparedBatches(
-      creates: Option[BatchSql],
-      exercises: Option[BatchSql],
-      archives: Option[BatchSql],
-  ) {
-    final def isEmpty: Boolean = creates.isEmpty && exercises.isEmpty && archives.isEmpty
-    final def foreach[U](f: BatchSql => U): Unit = {
-      creates.foreach(f)
-      exercises.foreach(f)
-      archives.foreach(f)
-    }
-  }
-
-  private final case class AccumulatingBatches(
-      creates: Vector[Vector[NamedParameter]],
-      exercises: Vector[Vector[NamedParameter]],
-      archives: Vector[Vector[NamedParameter]],
-  ) {
-
-    def addCreate(create: Vector[NamedParameter]): AccumulatingBatches =
-      copy(creates = creates :+ create)
-
-    def addExercise(exercise: Vector[NamedParameter]): AccumulatingBatches =
-      copy(exercises = exercises :+ exercise)
-
-    def addArchive(archive: Vector[NamedParameter]): AccumulatingBatches =
-      copy(archives = archives :+ archive)
-
-    private def prepareNonEmpty(
-        query: String,
-        params: Vector[Vector[NamedParameter]],
-    ): Option[BatchSql] =
-      if (params.nonEmpty) Some(BatchSql(query, params.head, params.tail: _*)) else None
-
-    def prepare: PreparedBatches =
-      new PreparedBatches(
-        prepareNonEmpty(insertCreate, creates),
-        prepareNonEmpty(insertExercise, exercises),
-        prepareNonEmpty(updateArchived, archives),
-      ) {}
-
-  }
-
-  private object AccumulatingBatches {
-    val empty: AccumulatingBatches = AccumulatingBatches(
-      creates = Vector.empty,
-      exercises = Vector.empty,
-      archives = Vector.empty,
-    )
-  }
-
-  /**
-    * @throws RuntimeException If a value cannot be serialized into an array of bytes
-    */
-  @throws[RuntimeException]
-  def prepareBatchInsert(
-      applicationId: Option[ApplicationId],
-      workflowId: Option[WorkflowId],
-      transactionId: TransactionId,
-      commandId: Option[CommandId],
-      submitter: Option[Party],
-      roots: Set[NodeId],
-      ledgerEffectiveTime: Date,
-      offset: Offset,
-      transaction: Transaction,
-  ): PreparedBatches =
-    transaction
-      .fold(AccumulatingBatches.empty) {
-        case (batches, (nodeId, node: Create)) =>
-          batches.addCreate(
-            create(
-              applicationId = applicationId,
-              workflowId = workflowId,
-              commandId = commandId,
-              transactionId = transactionId,
-              nodeId = nodeId,
-              submitter = submitter,
-              roots = roots,
-              ledgerEffectiveTime = ledgerEffectiveTime,
-              offset = offset,
-              create = node,
+    def toFlatTransaction(events: List[Entry[Event]]): Option[GetFlatTransactionResponse] = {
+      events.headOption.flatMap { first =>
+        val flatEvents =
+          TransactionConversion.removeTransient(events.iterator.map(_.event).toVector)
+        if (flatEvents.nonEmpty || first.commandId.nonEmpty)
+          Some(
+            GetFlatTransactionResponse(
+              transaction = Some(
+                ApiTransaction(
+                  transactionId = first.transactionId,
+                  commandId = first.commandId,
+                  effectiveAt =
+                    Some(Timestamp.of(seconds = first.ledgerEffectiveTime.getTime, nanos = 0)),
+                  workflowId = first.workflowId,
+                  offset = ApiOffset.toApiString(first.eventOffset),
+                  events = flatEvents,
+                )
+              )
             )
           )
-        case (batches, (nodeId, node: Exercise)) =>
-          val batchWithExercises =
-            batches.addExercise(
-              exercise(
-                applicationId = applicationId,
-                workflowId = workflowId,
-                commandId = commandId,
-                transactionId = transactionId,
-                nodeId = nodeId,
-                submitter = submitter,
-                roots = roots,
-                ledgerEffectiveTime = ledgerEffectiveTime,
-                offset = offset,
-                exercise = node,
-              )
-            )
-          if (node.consuming) {
-            batchWithExercises.addArchive(
-              archive(
-                contractId = node.targetCoid,
-                consumedAt = offset,
-              )
-            )
-          } else {
-            batchWithExercises
-          }
-        case (batches, _) =>
-          batches // ignore any event which is not a create or an exercise
+        else None
+
       }
-      .prepare
-
-  import SqlParser._
-
-  val createdEventParser: RowParser[RawFlatEvent] =
-    offset("event_offset") ~ str("transaction_id") ~ str("event_id") ~ str("contract_id") ~ date(
-      "ledger_effective_time") ~ str("template_package_id") ~ str("template_name") ~ get[Option[
-      String]]("command_id") ~ get[Option[String]]("workflow_id") ~ binaryStream("create_argument") ~ array[
-      String]("create_signatories") ~ array[String]("create_observers") ~ get[Option[String]](
-      "create_agreement_text") ~ get[Option[InputStream]]("create_key_value") ~ array[String](
-      "event_witnesses") map {
-      case eventOffset ~ transactionId ~ eventId ~ contractId ~ ledgerEffectiveTime ~ templatePackageId ~ templateName ~ commandId ~ workflowId ~ createArgument ~ createSignatories ~ createObservers ~ createAgreementText ~ createKeyValue ~ eventWitnesses =>
-        val QualifiedName(templateModuleName, templateEntityName) =
-          QualifiedName.assertFromString(templateName)
-        RawFlatEvent(
-          eventOffset = eventOffset,
-          transactionId = transactionId,
-          ledgerEffectiveTime = ledgerEffectiveTime,
-          commandId = commandId.getOrElse(""),
-          workflowId = workflowId.getOrElse(""),
-          event = Event(
-            Event.Event.Created(
-              CreatedEvent(
-                eventId = eventId,
-                contractId = contractId,
-                templateId = Some(
-                  Identifier(
-                    packageId = templatePackageId,
-                    moduleName = templateModuleName.dottedName,
-                    entityName = templateEntityName.dottedName,
-                  )),
-                contractKey = createKeyValue.map(
-                  key =>
-                    LfEngineToApi.assertOrRuntimeEx(
-                      failureContext = s"attempting to deserialize persisted key to value",
-                      LfEngineToApi
-                        .lfVersionedValueToApiValue(
-                          verbose = true,
-                          value = ValueSerializer.deserializeValue(key),
-                        ),
-                  )
-                ),
-                createArguments = Some(
-                  LfEngineToApi.assertOrRuntimeEx(
-                    failureContext =
-                      s"attempting to deserialize persisted create argument to record",
-                    LfEngineToApi
-                      .lfVersionedValueToApiRecord(
-                        verbose = true,
-                        recordValue = ValueSerializer.deserializeValue(createArgument),
-                      ),
-                  )
-                ),
-                witnessParties = eventWitnesses,
-                signatories = createSignatories,
-                observers = createObservers,
-                agreementText = createAgreementText,
-              )
-            )
-          )
-        )
     }
 
-  val archivedEventParser: RowParser[RawFlatEvent] =
-    offset("event_offset") ~ str("transaction_id") ~ str("event_id") ~ str("contract_id") ~ date(
-      "ledger_effective_time") ~ str("template_package_id") ~ str("template_name") ~ get[
-      Option[String]]("command_id") ~ get[Option[String]]("workflow_id") ~ array[String](
-      "event_witnesses") map {
-      case eventOffset ~ transactionId ~ eventId ~ contractId ~ ledgerEffectiveTime ~ templatePackageId ~ templateName ~ commandId ~ workflowId ~ eventWitnesses =>
-        val QualifiedName(templateModuleName, templateEntityName) =
-          QualifiedName.assertFromString(templateName)
-        RawFlatEvent(
-          eventOffset = eventOffset,
-          transactionId = transactionId,
-          ledgerEffectiveTime = ledgerEffectiveTime,
-          commandId = commandId.getOrElse(""),
-          workflowId = workflowId.getOrElse(""),
-          event = Event(
-            Event.Event.Archived(
-              ArchivedEvent(
-                eventId = eventId,
-                contractId = contractId,
-                templateId = Some(
-                  Identifier(
-                    packageId = templatePackageId,
-                    moduleName = templateModuleName.dottedName,
-                    entityName = templateEntityName.dottedName,
-                  )),
-                witnessParties = eventWitnesses,
-              )
-            )
-          )
-        )
-    }
-
-  val flatEventParser: RowParser[RawFlatEvent] =
-    createdEventParser | archivedEventParser
-
-  def prepareLookupFlatTransactionById(
-      transactionId: TransactionId,
-      requestingParties: Set[Party],
-  ): SimpleSql[Row] = {
-    val tx: String = transactionId
-    val ps: Set[String] = requestingParties.asInstanceOf[Set[String]]
-    SQL"""select
-            event_offset,
-            transaction_id,
-            ledger_effective_time,
-            case when submitter in ($ps) then command_id else '' end as command_id,
-            workflow_id,
-            participant_events.event_id,
-            contract_id,
-            template_package_id,
-            template_name,
-            create_argument,
-            create_signatories,
-            create_observers,
-            create_agreement_text,
-            create_key_value,
-            array_agg(event_witness) as event_witnesses
-          from participant_events
-          natural join participant_event_flat_transaction_witnesses
-          where transaction_id = $tx and event_witness in ($ps)
-          group by (
-            event_offset,
-            transaction_id,
-            ledger_effective_time,
-            command_id,
-            workflow_id,
-            participant_events.event_id,
-            contract_id,
-            template_package_id,
-            template_name,
-            create_argument,
-            create_signatories,
-            create_observers,
-            create_agreement_text,
-            create_key_value
-          )
-       """
   }
+
+  private type SharedRow =
+    Offset ~ String ~ String ~ String ~ Date ~ String ~ String ~ Option[String] ~ Option[String] ~ Array[
+      String]
+  private val sharedRow: RowParser[SharedRow] =
+    offset("event_offset") ~
+      str("transaction_id") ~
+      str("event_id") ~
+      str("contract_id") ~
+      date("ledger_effective_time") ~
+      str("template_package_id") ~
+      str("template_name") ~
+      get[Option[String]]("command_id") ~
+      get[Option[String]]("workflow_id") ~
+      array[String]("event_witnesses")
+
+  protected type CreatedEventRow =
+    SharedRow ~ InputStream ~ Array[String] ~ Array[String] ~ Option[String] ~ Option[InputStream]
+  protected val createdEventRow: RowParser[CreatedEventRow] =
+    sharedRow ~
+      binaryStream("create_argument") ~
+      array[String]("create_signatories") ~
+      array[String]("create_observers") ~
+      get[Option[String]]("create_agreement_text") ~
+      get[Option[InputStream]]("create_key_value")
+
+  protected type ArchiveEventRow = SharedRow
+  protected val archivedEventRow: RowParser[ArchiveEventRow] = sharedRow
+
+  private def templateId(
+      templatePackageId: String,
+      templateName: String,
+  ): Identifier = {
+    val QualifiedName(templateModuleName, templateEntityName) =
+      QualifiedName.assertFromString(templateName)
+    Identifier(
+      packageId = templatePackageId,
+      moduleName = templateModuleName.dottedName,
+      entityName = templateEntityName.dottedName,
+    )
+  }
+
+  protected def createdEvent(
+      eventId: String,
+      contractId: String,
+      templatePackageId: String,
+      templateName: String,
+      createArgument: InputStream,
+      createSignatories: Array[String],
+      createObservers: Array[String],
+      createAgreementText: Option[String],
+      createKeyValue: Option[InputStream],
+      eventWitnesses: Array[String],
+  ): CreatedEvent =
+    CreatedEvent(
+      eventId = eventId,
+      contractId = contractId,
+      templateId = Some(templateId(templatePackageId, templateName)),
+      contractKey = createKeyValue.map(
+        key =>
+          LfEngineToApi.assertOrRuntimeEx(
+            failureContext = s"attempting to deserialize persisted key to value",
+            LfEngineToApi
+              .lfVersionedValueToApiValue(
+                verbose = true,
+                value = deserialize(key),
+              ),
+        )
+      ),
+      createArguments = Some(
+        LfEngineToApi.assertOrRuntimeEx(
+          failureContext = s"attempting to deserialize persisted create argument to record",
+          LfEngineToApi
+            .lfVersionedValueToApiRecord(
+              verbose = true,
+              recordValue = deserialize(createArgument),
+            ),
+        )
+      ),
+      witnessParties = eventWitnesses,
+      signatories = createSignatories,
+      observers = createObservers,
+      agreementText = createAgreementText,
+    )
+
+  protected def archivedEvent(
+      eventId: String,
+      contractId: String,
+      templatePackageId: String,
+      templateName: String,
+      eventWitnesses: Array[String],
+  ): ArchivedEvent =
+    ArchivedEvent(
+      eventId = eventId,
+      contractId = contractId,
+      templateId = Some(templateId(templatePackageId, templateName)),
+      witnessParties = eventWitnesses,
+    )
 
 }
